+++ src/virtualenv.py
@@ -1112,7 +1112,13 @@ def change_prefix(filename, dst_prefix):
 
 
 def copy_required_modules(dst_prefix, symlink):
-    import imp
+    import warnings
+
+    with warnings.catch_warnings():
+        # Ignore deprecation of the imp module
+        # TODO: do not use deprecated imp module
+        warnings.simplefilter("ignore")
+        import imp
 
     for modname in REQUIRED_MODULES:
         if modname in sys.builtin_module_names: