+++ docs/changes.rst
@@ -12,7 +12,7 @@ Release History
 * ``activate.csh`` does not use basename and handles newlines :issue:`1200`
 * fix failure to copy on platforms that use lib64 :issue:`1189`
 * enable tab-completion in the interactive interpreter by default, thanks to a new ``sys.__interactivehook__`` on Python 3 :issue:`967`
-
+* suppress warning of usage of the deprecated ``imp`` module :issue:`1238`
 
 16.0.0 (2018-05-16)
 -------------------